# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VenueNav API
  version: 1.0.0
  description: |
    REST API for multi-tenant indoor event maps: upload, processing jobs, graph/shops, routing, and search.
    All mutating admin routes require appropriate organization role (owner, editor).
servers:
  - url: https://api.venuenav.example.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrgId:
      name: org_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EventId:
      name: event_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MapId:
      name: map_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    MapSummary:
      type: object
      required: [id, name, status, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        published_version:
          type: integer
          nullable: true
        updated_at:
          type: string
          format: date-time
    MapVersion:
      type: object
      required: [id, version, width, height, unit, created_at]
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
        width:
          type: number
        height:
          type: number
        unit:
          type: string
          enum: [pixel, meter]
        srid:
          type: integer
          description: PostGIS SRID for stored geometries; 0 or custom for local.
    GraphNode:
      type: object
      required: [id, x, y, type]
      properties:
        id:
          type: string
        x:
          type: number
        y:
          type: number
        type:
          type: string
          enum: [intersection, entrance, exit, shop_entry, generic]
        label:
          type: string
          nullable: true
    GraphEdge:
      type: object
      required: [from, to, weight]
      properties:
        from:
          type: string
        to:
          type: string
        weight:
          type: number
          description: Cost (distance or time).
        edge_id:
          type: string
          format: uuid
          description: Server graph_edge id (for live overrides and admin tools).
...
    Shop:
      type: object
      required: [id, name, location_node]
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          nullable: true
        location_node:
          type: string
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
    MapPayload:
      type: object
      required: [map_id, map_version_id, nodes, edges, shops]
      properties:
        map_id:
          type: string
        map_version_id:
          type: string
        width:
          type: number
        height:
          type: number
        raster_url:
          type: string
          format: uri
          description: Canonical background image for rendering.
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/GraphNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/GraphEdge'
        shops:
          type: array
          items:
            $ref: '#/components/schemas/Shop'
...
    ProcessingJob:
      type: object
      required: [id, map_id, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        map_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, rasterizing, detecting, ocr, completed, failed, awaiting_review]
        progress:
          type: number
          minimum: 0
          maximum: 1
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
security:
  - bearerAuth: []
paths:
...
  /organizations/{org_id}/events/{event_id}/maps:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Maps]
      summary: List maps for an event
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MapSummary'
    post:
      tags: [Maps]
      summary: Create map record (then upload file via pre-signed URL)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, venue_id]
              properties:
                name: { type: string }
                venue_id: { type: string, format: uuid }
      responses: { '201': { description: Created } }
  /organizations/{org_id}/maps/{map_id}/upload:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/MapId'
    post:
      tags: [Maps]
      summary: Issue pre-signed URL(s) for PDF (and optional page index)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content_type:
                  type: string
                  default: application/pdf
                byte_size:
                  type: integer
      responses:
        '200':
          description: Upload instructions
          content:
            application/json:
              schema:
                type: object
                properties:
                  upload_url: { type: string, format: uri }
                  asset_id: { type: string, format: uuid }
                  headers:
                    type: object
                    additionalProperties: { type: string }
  /organizations/{org_id}/maps/{map_id}/process:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/MapId'
    post:
      tags: [Processing]
      summary: Start or restart map processing after PDF uploaded
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                asset_id: { type: string, format: uuid }
                options:
                  type: object
                  description: DPI, page, segmentation model, etc.
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingJob'
  /organizations/{org_id}/maps/{map_id}/jobs/{job_id}:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/MapId'
      - name: job_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Processing]
      summary: Get job status
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingJob'
  /organizations/{org_id}/maps/{map_id}/graph:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/MapId'
    get:
      tags: [Graph]
      summary: Get draft or published graph (query param)
      parameters:
        - name: state
          in: query
          schema:
            type: string
            enum: [draft, published]
            default: draft
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapPayload'
    put:
      tags: [Graph]
      summary: Replace draft graph (bulk save from admin editor)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MapPayload'
      responses: { '200': { description: OK } }
  /organizations/{org_id}/maps/{map_id}/publish:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/MapId'
    post:
      tags: [Graph]
      summary: Publish draft as new immutable version
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapVersion'
...
  /public/maps/{map_id}/payload:
    get:
      tags: [Maps]
      security: []
      summary: Full published graph + shops for client (cacheable, ETag)
      parameters:
        - name: map_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: version
          in: query
          schema: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapPayload'
...
  /organizations/{org_id}/maps/{map_id}/shops/{shop_id}:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/MapId'
      - name: shop_id
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Shops]
      summary: Update shop metadata and tags
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shop'
      responses: { '200': { description: OK } }